Separate inquirer script from import.js
